- Remove old workaround on template conditionnal selection - Rewrite uwp switch - Fix winsock redefinition - Add windows switch on headers - Add static libs in FindSoci search - Fix cast error

+++ tester/tester.c
@@ -1988,7 +1988,7 @@ void notify_presence_received_for_uri_or_tel(LinphoneCore *lc, LinphoneFriend *l
 }
 
 void _check_friend_result_list(LinphoneCore *lc, const bctbx_list_t *resultList, const unsigned int index, const char* uri, const char* phone) {
-	if (index >= bctbx_list_size(resultList)) {
+	if (index >= (unsigned int)bctbx_list_size(resultList)) {
 		ms_error("Attempt to access result to an outbound index");
 		return;
 	}
